+++ src/component/ProfileTrainer/Actu/index.jsx
@@ -1,23 +1,10 @@
-import CheckCircleOutlineIcon from "@mui/icons-material/CheckCircleOutline";
-import ErrorOutlineOutlinedIcon from "@mui/icons-material/ErrorOutlineOutlined";
-import HighlightOffIcon from "@mui/icons-material/HighlightOff";
-import PictureAsPdfIcon from "@mui/icons-material/PictureAsPdf";
 import SendIcon from "@mui/icons-material/Send";
 import { Box, Button, Modal, TextField } from "@mui/material";
 import { DataGrid } from "@mui/x-data-grid";
-import React, { useCallback, useEffect, useState } from "react";
-import { request } from "../../../core/api/request";
+import React from "react";
 import "./styles.css";
 import styles from "./styles.module.css";
-
-const buttonStyles = (disable) => ({
-  minWidth: "10px",
-  width: "70px",
-  cursor: disable ? "not-allowed" : "pointer",
-  "&.Mui-disabled": disable && {
-    pointerEvents: "auto",
-  },
-});
+import useNotifTrainer from "./useNotifTrainer";
 
 const modalStyle = {
   position: "absolute",
@@ -26,7 +13,7 @@ const modalStyle = {
   transform: "translate(-50%, -50%)",
   bgcolor: "background.paper",
   borderRadius: 2,
-  boxShadow: 24,
+  boxShadow: 100,
   p: 2,
   width: "30rem",
   height: "20rem",
@@ -34,187 +21,17 @@ const modalStyle = {
 };
 
 const Actu = () => {
-  const [openChange, setOpenChange] = useState(false);
-  const [valueComment, setValueComment] = useState({
-    Creator: "Me",
-    Comment: "",
-  });
-  const [data, setData] = useState([]);
-  const [comments, setComments] = useState([]);
-  const [currentElement, setCurrentElement] = useState("");
-
-  const handleChange = (e) => {
-    setValueComment((prev) => ({ ...prev, Comment: e.target.value }));
-  };
-
-  const getNotifTrainer = useCallback(() => {
-    request.read("Trainer/GetNotifTrainer").then((data) => {
-      setData(data.data);
-    });
-  }, []);
-
-  useEffect(() => getNotifTrainer(), [getNotifTrainer]);
-
-  const handleUpdate = async (id, status, price) => {
-    const jsonData = {
-      data: { id, reponseFormateur: status, prixFormateur: price },
-    };
-    request.create("Trainer/UpdateNotifTrainer", jsonData).then(() => {
-      // window.location.reload(true);
-    });
-  };
-  const renderResponseButton = (params) => {
-    const { statusMandat, reponseFormateur, id, TJ } = params.row;
-
-    let isDisabled = false;
-
-    let opacityConfirm = "20%";
-    let opacityReject = "20%";
-    let opacityWaiting = "100%";
-
-    if (["confirmed", "closed"].includes(statusMandat)) {
-      isDisabled = true;
-      opacityConfirm = statusMandat === "confirmed" ? "100%" : "20%";
-    }
-
-    if (reponseFormateur === "confirmed") {
-      isDisabled = true;
-      opacityConfirm = "100%";
-    }
-
-    if (reponseFormateur === "rejected") {
-      isDisabled = true;
-      opacityReject = "100%";
-    }
-
-    return (
-      <>
-        <Button
-          sx={buttonStyles(isDisabled)}
-          disabled={isDisabled}
-          onClick={() => handleUpdate(id, "confirmed", TJ)}
-        >
-          <CheckCircleOutlineIcon
-            sx={{ color: "green", opacity: opacityConfirm, width: "50px" }}
-          />
-        </Button>
-
-        <Button
-          sx={buttonStyles(isDisabled)}
-          disabled={isDisabled}
-          onClick={() => handleUpdate(id, "rejected")}
-        >
-          <HighlightOffIcon
-            sx={{ color: "red", opacity: opacityReject, width: "50px" }}
-          />
-        </Button>
-
-        <Button
-          sx={buttonStyles(isDisabled)}
-          disabled={isDisabled}
-          onClick={() => handleUpdate(id, "waiting")}
-        >
-          <ErrorOutlineOutlinedIcon
-            sx={{ color: "#ff9100", opacity: opacityWaiting, width: "50px" }}
-          />
-        </Button>
-      </>
-    );
-  };
-
-  const renderBCButton = (params) => (
-    <Button
-      sx={buttonStyles(params.row.statusMandat !== "confirmed")}
-      disabled={params.row.statusMandat !== "confirmed"}
-    >
-      <PictureAsPdfIcon
-        sx={{
-          color:
-            params.row.statusMandat === "confirmed" ? "#B33030" : "#7C99AC",
-        }}
-      />
-    </Button>
-  );
-
-  const renderCommentButton = (params) => (
-    <Button
-      sx={{ color: "red" }}
-      onClick={() => {
-        setCurrentElement(params.row.id);
-        setComments(params.row.comments);
-        setOpenChange(true);
-      }}
-    >
-      Explore
-    </Button>
-  );
-
-  const columns = [
-    { field: "subject", headerName: "Subject", width: 140 },
-    { field: "date", headerName: "Date", width: 150 },
-    {
-      field: "TJ",
-      headerName: "TJ (Frais inclus)",
-      width: 150,
-      editable: true,
-    },
-    { field: "statusMandat", headerName: "Status Mandat", width: 150 },
-    { field: "NB", headerName: "NB d'inscrit", width: 130 },
-    {
-      field: "RF",
-      headerName: "Réponse Formateur",
-      width: 190,
-      renderCell: renderResponseButton,
-    },
-    {
-      field: "BC",
-      headerName: "Bon de commande",
-      width: 190,
-      renderCell: renderBCButton,
-    },
-    {
-      field: "commentaire",
-      headerName: "Commentaire",
-      width: 190,
-      renderCell: renderCommentButton,
-    },
-  ];
-
-  const rows = data.map((row) => ({
-    id: row._id,
-    subject: row.courseCertif,
-    date: `${new Date(row.courseDate[0]).getDate()}/${
-      new Date(row.courseDate[0]).getMonth() + 1
-    } -> ${new Date(row.courseDate[1]).getDate()}/${
-      new Date(row.courseDate[1]).getMonth() + 1
-    } - ${new Date(row.courseDate[1]).getFullYear()}`,
-    TJ: row.prixFormateur,
-    statusMandat: row.StatusMandate,
-    NB: row.nbInscrit,
-    reponseFormateur: row.reponseFormateur,
-    comments: row.comments,
-  }));
-
-  const columnsComments = [
-    { field: "Creator", headerName: "Creator", width: 140 },
-    { field: "Comment", headerName: "Comment", width: 640 },
-  ];
-
-  const rowsComments = comments.map((row, index) => ({
-    id: index,
-    Creator: row.Creator,
-    Comment: row.Comment,
-  }));
-
-  const SendComment = async () => {
-    const jsonData = {
-      data: { id: currentElement, comments: valueComment },
-    };
-    request.create("Trainer/UpdateCommentsNotifTrainer", jsonData).then(() => {
-      getNotifTrainer();
-      setOpenChange(false);
-    });
-  };
+  const {
+    openChange,
+    setOpenChange,
+    valueComment,
+    handleChange,
+    rows,
+    columns,
+    columnsComments,
+    rowsComments,
+    SendComment,
+  } = useNotifTrainer();
 
   return (
     <div className={styles.leftSectionProfile}>
